Handling Patrolling
