+++ Assets/_BForBoss/_Core/Scripts/Environment.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6655be6f4677c064899ddc386e7ec41e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
