+++ Assets/_BForBoss/_Core/Scripts/Environment/Patrol.cs
@@ -1,54 +1,16 @@
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using ECM2.Components;
-using Perigon.Utility;
-using PerigonGames;
 using UnityEngine;
 
 namespace BForBoss
 {
-    public class PatrolBehaviour : PlatformMovement
-    {
-        [SerializeField] private List<GameObject> _patrolPosition = new List<GameObject>();
-        [SerializeField] private float _speed = 1;
-        [SerializeField] private float _waitTimeOnArrival = 3;
-        
-        private Patrol _patrol = null;
-
-        public void CleanUp()
-        {
-            _patrol.CleanUp();
-        }
-
-        public void Reset()
-        {
-            _patrol.Reset();
-        }
-
-        private void Awake()
-        {
-            if (_patrolPosition.IsNullOrEmpty())
-            {
-                PanicHelper.Panic(new Exception("List of patrol position is empty or null for Patrol Behaviour"));
-            }
-            _patrol = new Patrol(_patrolPosition.Select(x => x.transform.position).ToArray(), _speed, _waitTimeOnArrival);    
-            transform.position = _patrolPosition[0].transform.position;
-        }
-
-        protected override void OnMove()
-        {
-            position = _patrol.MoveTowards(transform.position, Time.deltaTime);
-        }
-    }
-    
-    public class Patrol
+     public class Patrol
     {
         private enum PatrolState
         {
             Moving,
             Waiting
         }
+        
         private readonly Queue<Vector3> _queueOfLocations = new Queue<Vector3>();
         private readonly float _speed = 0;
         private readonly float _waitTimeOnArrival = 0;
@@ -74,7 +36,6 @@ namespace BForBoss
         {
             if (_state == PatrolState.Waiting && NeedToWait)
             {
-                Debug.Log("waiting");
                 UpdateWaitTime(time);
                 return position;
             }
@@ -82,7 +43,6 @@ namespace BForBoss
             var nextPosition = Vector3.MoveTowards(position, _currentDestination, _speed * time);
             if (nextPosition == _currentDestination)
             {
-                Debug.Log("Got next destination: " + _currentDestination);
                 _state = PatrolState.Waiting;
                 GetNextDestination();
             }
@@ -95,7 +55,6 @@ namespace BForBoss
             _elapsedWaitTime += time;
             if (_elapsedWaitTime > _waitTimeOnArrival)
             {
-                Debug.Log("Finished Waiting");
                 _elapsedWaitTime = 0;
                 _state = PatrolState.Moving;
             }
