+++ Assets/_BForBoss/_Core/Scripts/Environment/Patrol.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b4c91aa1adb83204c89a6203f82194e8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
